Eliminate a few compiler warnings

+++ FreeRTOS_Sockets.c
@@ -4517,8 +4517,8 @@ BaseType_t xSocketValid( Socket_t xSocket )
                 {
                     /* Using function "snprintf". */
                     const int32_t copied_len = snprintf( ucChildText, sizeof( ucChildText ), " %d/%d",
-                                                         ( int32_t ) pxSocket->u.xTCP.usChildCount,
-                                                         ( int32_t ) pxSocket->u.xTCP.usBacklog );
+                                                         pxSocket->u.xTCP.usChildCount,
+                                                         pxSocket->u.xTCP.usBacklog );
                     ( void ) copied_len;
                     /* These should never evaluate to false since the buffers are both shorter than 5-6 characters (<=65535) */
                     configASSERT( copied_len >= 0 );
